component props fixed
