+++ src/components/pages/TripList/AddTripPlan.module.scss
@@ -1,26 +1,11 @@
 .root {
   width: 100vw;
-  .loading {
-    position: absolute;
-    top: 30%;
-    left: 50%;
-    font-size: x-large;
-    font-weight: 800;
-    color: #a9d159;
-  }
-  .none_text {
-    position: absolute;
-    top: 45%;
-    left: 50%;
-    font-size: medium;
-    font-weight: 800;
-    color: #a9d159;
-  }
   .plan_name {
     width: 60%;
     position: absolute;
     left: 15%;
-    margin: 20px;
+    margin-bottom: 10px;
+    padding-bottom: 10px;
     font-size: x-large;
     background-color: white;
     border: 1px solid #a9d159;
