+++ src/pages/Login/Login.module.scss
@@ -3,6 +3,8 @@
   position: relative;
   width: 100vw;
   min-height: 100vh;
+  background: -webkit-linear-gradient(top, rgba(253, 219, 146, 1), rgba(209, 253, 254, 1));
+  overflow: hidden;
   .register_image {
     position: absolute;
     background-size: cover;
@@ -13,12 +15,13 @@
     display: flex;
     position: relative;
     justify-content: center;
-    background-color: white;
+    background-color: #f0fff0;
     position: absolute;
     top: 140px;
     left: 40%;
     width: 500px;
     height: 700px;
+    border-radius: 20px;
     .title {
       justify-content: center;
       position: absolute;
@@ -33,7 +36,6 @@
       top: 35%;
       left: 50%;
       transform: translate(-50%, -50%);
-      background-color: white;
       color: white;
       width: 75%;
       margin-bottom: 5px;
@@ -44,7 +46,6 @@
       top: 45%;
       left: 50%;
       transform: translate(-50%, -50%);
-      background-color: white;
       color: white;
       width: 75%;
     }
@@ -54,7 +55,6 @@
       top: 55%;
       left: 50%;
       transform: translate(-50%, -50%);
-      background-color: white;
       color: white;
       width: 75%;
     }
